fix warnings in vs2010

+++ pnas04/src/cell.cpp
@@ -70,7 +70,7 @@ void Cell::mut_deg_prot () {
 	
 	if(!num1) return;	//	no protein
 
-    srand(time(NULL));
+    srand((unsigned int)time(NULL));
     int opIndex = indice[rand()%num1];//opIndex contains a certain degradation reaction
 
     //  modify its degradation rate
@@ -86,7 +86,7 @@ void Cell::mut_kin_const () {
 	
 	//	a kinetic constant of one reaction is modified
 	
-	srand(time(NULL));
+	srand((unsigned int)time(NULL));
 	Reaction* currR = rlist[rand()%rlist.size()];
 
 	if((double)rand()/RAND_MAX<=0.5 || !currR->isReversible()) {
@@ -157,7 +157,7 @@ void Cell::mut_add_regu () {
 	if(!num1 || !num2) return; // no gene or gene/protein complex
 
 	//	sample random number
-    srand(time(NULL));
+    srand((unsigned int)time(NULL));
     int opIndex1 = protIndice[rand()%num1];//contain index of a protein
     int opIndex2 = cplxIndice[rand()%num2];//contain index of a gene or a gene/protein complex
 	Node* exGene = nodes[opIndex2]->extractFirstGene();
@@ -206,7 +206,7 @@ void Cell::mut_add_regu () {
 void Cell::mut_add_postmod () {
 	//	a post modification is add
 	
-    srand(time(NULL));
+    srand((unsigned int)time(NULL));
 	if((double)rand()/RAND_MAX <= 0.5) {//	a single protein case
 
 
@@ -276,7 +276,7 @@ void Cell::mut_add_postmod () {
 
 //overall mutation method
 void Cell::mutation(){
-    srand(time(NULL));
+    srand((unsigned int)time(NULL));
     if (rand() < RAND_MAX*PROB1) {
         mut_deg_prot();
     }
@@ -332,8 +332,6 @@ void runge_kutta(double data[][MAXSTEPS],double (*odes[])(double y[],double x),i
     }
     delete y, tempY;
 }
-
-
 
 
 /*get score using the sfunc as score function and change its own currScore member
